Update hidden_test.html
Dragging and Moving Cards:
Cards are now draggable from their initial locations in the boxes.
Dragging an individual card will move it into the play field. Once moved, it condenses as part of the committed discard area.
Committed Cards:
When cards are played, they move to the "discard" area as face-down cards, then condense (by stacking) to show a counter for the number of cards in that pile.
Card Update Logic:
The Play Cards! button will increment through a series of cards and update the current card to be displayed.
'BS!' Button:
The button checks if the latest card is consistent with the play order.